Assignments

+++ src/com/bridgelabz/day2/Springseason.java
@@ -3,8 +3,8 @@ package com.bridgelabz.day2;
 public class Springseason {
 
 	public static void main(String[] args) {
-		int m=Integer.parseInt(args[0]);
-		int d=Integer.parseInt(args[1]);
+		byte m=Byte.parseByte(args[0]);
+		byte d=Byte.parseByte(args[1]);
 		boolean b=false;
 
 		if(m==3){
@@ -29,7 +29,6 @@ public class Springseason {
 		}
 		else
 		System.out.println(b);
-
 	}
 
 }
